pyproject.toml
[tool.poetry]
name = "pyramid_temporal"
version = "0.0.1"
description = "This project bind the config and transaction from pyramid to temporal"
authors = ["Robbin dev team <fdev@robbin.com.br>"]
repository = "https://github.com/tomas_correa/pyramid-temporal"
documentation = "https://tomas_correa.github.io/pyramid-temporal/"
readme = "README.md"
packages = [
  {include = "pyramid_temporal"}
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
temporalio = "^1.15.0"
transaction = "^5.0"
zope-interface = "^7.2"
pyramid = "^2.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^4.0.0"
deptry = "^0.6.4"
mypy = "^0.981"
pre-commit = "^2.20.0"
tox = "^3.25.1"
sqlalchemy = "^2.0.43"
alembic = "^1.16.4"
psycopg2-binary = "^2.9.10"
pyramid-tm = "^2.6"
testing-postgresql = "^1.3.0"
webtest = "^3.0.6"
pytest-asyncio = "^1.1.0"
zope-sqlalchemy = "^3.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocstrings = "^0.29.1"
mkdocstrings-python = "^1.16.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py37']
preview = true

[tool.mypy]
files = ["pyramid_temporal"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.ruff]
target-version = "py37"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["pyramid_temporal"]


